requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.1
click==8.1.7
Werkzeug==3.0.1 